requirements.txt
2to3==1.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
RPi.GPIO==0.7.0
six==1.14.0
smbus==1.1.post2
